use short for PathSegment
Stackallocing the array of PathSegments is a significant cost in
routing's plaintext implementation due to zero-initialization. This
change cuts the size in half. This also has the benefit of fitting
PathSegment into a register on 32bit.

Note that profiling data in the PR is from 64bit, so this should be an
improvement regardless of bitness.

+++ src/Microsoft.AspNetCore.Routing/Matching/FastPathTokenizer.cs
@@ -38,7 +38,7 @@ namespace Microsoft.AspNetCore.Routing.Matching
             var span = path.AsSpan(start);
             while ((end = span.IndexOf('/')) >= 0 && count < segments.Length)
             {
-                segments[count++] = new PathSegment(start, end);
+                segments[count++] = new PathSegment((short)start, (short)end);
                 start += end + 1; // resume search after the current character
                 span = path.AsSpan(start);
             }
@@ -47,7 +47,7 @@ namespace Microsoft.AspNetCore.Routing.Matching
             var length = span.Length;
             if (length > 0 && count < segments.Length)
             {
-                segments[count++] = new PathSegment(start, length);
+                segments[count++] = new PathSegment((short)start, (short)length);
             }
 
             return count;
